+++ RELEASE-NOTES.md
@@ -4,6 +4,7 @@
 * #621 Problem with timestamp formatting
 * #622 Upgrade Jackson to 2.14.2 
 * #624 Optimization: "relatedViaQueryWithParentIds" strategy when limit is involved
+* #629 Upgrade to Cayenne 4.2 GA
 
 ## Release 5.0.M15